requirements.txt
fastapi
uvicorn[standard]
python-multipart
PyMuPDF==1.25.3
transformers
torch
Pillow
pdf2image
python-docx
aiofiles
